+++ src/Prdb.Fab.Infrastructure/Persistence/Migrations/FabDbContextModelSnapshot.cs
@@ -974,8 +974,8 @@ namespace Prdb.Fab.Infrastructure.Persistence.Migrations
                             DeleteLeftovers = false,
                             IndexersSkipped = false,
                             OnboardingStep = "Password",
-                            ReportConfirmedAssignments = false,
-                            ReportFulfilments = false,
+                            ReportConfirmedAssignments = true,
+                            ReportFulfilments = true,
                             RetryBudget = 3,
                             SabnzbdSkipped = false
                         });